refactor: EmptyPage
